requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
Django==3.2.5
django-bootstrap-modal-forms==2.2.0
django-braces==1.14.0
django-crispy-forms==1.9.0
django-gravatar2==1.4.4
django-js-asset==1.2.2
django-ledger==0.4.3.3
django-model-utils==4.0.0
django-mptt==0.12.0
django-star-ratings==0.9.1
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
drf-dynamic-fields==0.3.1
Faker==8.10.1
humanize==3.11.0
idna==2.10
importlib-metadata==4.6.1
jalali==1.2
jsonfield==3.1.0
jsonschema==3.2.0
Markdown==3.3.4
msal==1.5.0
numpy==1.21.0
ofxtools==0.8.22
pandas==1.2.5
persiantools==2.1.1
Pillow==7.0.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.18.0
python-dateutil==2.8.1
python-decouple==3.3
python-power-bi==0.1.1
pytz==2019.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
text-unidecode==1.3
typing-extensions==3.10.0.0
urllib3==1.25.11
zipp==3.5.0
